frontend/requirements.txt
streamlit
python-dotenv==1.0.1